added generate random password function
