+++ Navigation/Extensions/String + Chars.swift
@@ -1,0 +1,24 @@
+//
+//  String + Chars.swift
+//  Navigation
+//
+//  Created by Dima Gorbachev on 12.08.2021.
+//  Copyright © 2021 Artem Novichkov. All rights reserved.
+//
+
+import Foundation
+
+extension String {
+    var digits:      String { return "0123456789" }
+    var lowercase:   String { return "abcdefghijklmnopqrstuvwxyz" }
+    var uppercase:   String { return "ABCDEFGHIJKLMNOPQRSTUVWXYZ" }
+    var punctuation: String { return "!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~" }
+    var letters:     String { return lowercase + uppercase }
+    var printable:   String { return digits + letters + punctuation }
+    
+    mutating func replace(at index: Int, with character: Character) {
+        var stringArray = Array(self)
+        stringArray[index] = character
+        self = String(stringArray)
+    }
+}
